% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcuminc.R
\name{ggcuminc}
\alias{ggcuminc}
\title{Lisa's Cumulative Incidence Plots}
\usage{
ggcuminc(msfit, groups = NA, grlabs = NA, grname = " ",
  events = NA, evlabs = NA, evtitle = "Event", perc = FALSE,
  xlab = "Time", ylab = "Probability", xlim = NA, ylim = c(0, 1),
  xby = NA, yby = 0.1, xbrlabs = NA, stack = "events",
  state0 = "Healthy", cuminc.col = NA, cuminc.lty = NA,
  cuminc.size = 1.5, step = TRUE, main = "", graystest = FALSE,
  df = NA, timevar = NA, eventvar = NA)
}
\arguments{
\item{msfit}{survfit object of type "mstate" (REQUIRED).}

\item{groups}{Character vector for groups as listed in dataset. Default = NA (no groups).}

\item{grlabs}{Character vector for group names, must be in same order as groups. Default = NA (no groups or use group levels from dataset).}

\item{grname}{Character label for group legend. Default = " ".}

\item{events}{Character vector for event names (including state 0).}

\item{evlabs}{Character vector for event labels.}

\item{evtitle}{Character label for event legend. Default = "Events".}

\item{perc}{Logical indicator to show y axis as percentages. Default = FALSE.}

\item{xlab}{Character label for x axis. Default = "Time".}

\item{ylab}{Character label for y axis. Default = "Probability".}

\item{xlim}{Numeric vector of length 2 for x axis min and max. Default = NA, using 0 to data maximum.}

\item{ylim}{Numeric vector of length 2 for y axis min and max. Default = c(0, 1).}

\item{xby}{Number for x axis major tick marks. Default = NA (ggplot selected).}

\item{yby}{Numeric for y axis major tick marks. Default = 0.1.}

\item{xbrlabs}{Vector for alternative x axis labels. Default = NA.}

\item{stack}{Character to indicate stacking events "events" or groups "groups". Default = "events".}

\item{state0}{Character label for state 0. Default = NA (do not display estimates for state 0).}

\item{cuminc.col}{Character vector for color of shaded cumulative incidence bands. Default = default colors.}

\item{cuminc.lty}{Numeric vector for line type of cumulative incidence lines. Default = 1.}

\item{cuminc.size}{Number for size of cumulative incidence lines. Default = 1.}

\item{step}{Logical indicator to use step function or not. Default = TRUE.}

\item{main}{Character label for plot title. Default = " ".}

\item{graystest}{Logical indicator to provide Gray's test p-values (Default = FALSE). If TRUE, you must provide df, timevar, and eventvar.}

\item{df}{Dataframe to use for Gray's test. Default = NA.}

\item{timevar}{Character name of time to event variable. Default = NA.}

\item{eventvar}{Character nane of event variable. Default = NA.}
}
\description{
This function creates cumulative incidence plots from a survfit object of type "mstate" using gglot2 graphics.
It returns a ggplot2 object and prints the resulting plot.
}
\keyword{Lisa}
\keyword{cmprsk}
\keyword{consulting}
\keyword{cumulative}
\keyword{incidence}
\keyword{mstate}
\keyword{plot}
\keyword{summary}
\keyword{survival}
