% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medtab.R
\name{medtab}
\alias{medtab}
\title{Lisa's Median Survival Table}
\usage{
medtab(sfit, groups = NA, grlabs = NA, grcolname = "Group",
  med.dec = 2, medlab = "Median survival time", printorig = TRUE,
  kable = TRUE, htmlTable = FALSE)
}
\arguments{
\item{sfit}{Survfit object (REQUIRED).}

\item{groups}{Character vector for groups as listed in dataset. Default = NA (no groups).}

\item{grlabs}{Character vector for group names, must be in same order as groups. Default = NA (no groups or use group levels from dataset).}

\item{grcolname}{Character label for group column. Default = "Group".}

\item{med.dec}{Number of decimal places to report for median survival time. Default = 2.}

\item{medlab}{Character label for median and CI columns. Default = "Median survival time".}

\item{printorig}{Logical indicator to print original summary of survfit object for checking purposes. Default = TRUE.}

\item{kable}{Logical. Indicator to use kable to display table. Default = TRUE.}

\item{htmlTable}{Logical. Indicator to use htmlTable package to display table instead of kable Default = FALSE.}
}
\description{
This function creates a summary table from a survfit object including median survival estimates with CIs.
It returns a dataframe and prints a table using kable or htmlTable.
}
\keyword{Lisa}
\keyword{consulting}
\keyword{summary}
\keyword{survival}
\keyword{table}
