% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtab.R
\name{survtab}
\alias{survtab}
\title{Lisa's Survival Summary Table}
\usage{
survtab(sfit, times, timelab = "Time", groups = NA, grlabs = NA,
  surv.dec = 2, perc = FALSE, color = "#EEEEEE", printorig = TRUE)
}
\arguments{
\item{sfit}{Survfit object (REQUIRED).}

\item{times}{Numeric vector of times for survival estimates (REQUIRED).}

\item{timelab}{Character label for time-points. Default = "Time".}

\item{groups}{Character vector for groups as listed in dataset. Default = NA (no groups).}

\item{grlabs}{Character vector for group names, must be in same order as groups. Default = NA (no groups or use group levels from dataset).}

\item{surv.dec}{Number of decimal places to report for survival estimates. Default = 2.}

\item{perc}{Logical indicator to report survival estimates as percentages. Default = FALSE.}

\item{color}{Character Hex color to use for htmlTable striping. Default = "#EEEEEE" (light grey).}

\item{printorig}{Logical indicator to print original summary of survfit object for checking purposes. Default = TRUE.}
}
\description{
This function creates a summary table from a survfit object including survival estimates with CIs, number of events, and number at risk.
It returns a dataframe and prints an html table by default for use in R markdown documents.
}
\keyword{Lisa}
\keyword{consulting}
\keyword{summary}
\keyword{survival}
\keyword{table}
