% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citab.R
\name{citab}
\alias{citab}
\title{Lisa's Cumulative Incidence Survival Table}
\usage{
citab(msfit, times, timelab = "Time", groups = NA, grlabs = NA,
  state0 = NA, ci.dec = 2, perc = FALSE, color = "#EEEEEE",
  printorig = TRUE)
}
\arguments{
\item{msfit}{survfit object of type "mstate" (REQUIRED).}

\item{times}{Numeric vector of times for cumulative incidence estimates (REQUIRED).}

\item{timelab}{Character label for time. Default = "Time".}

\item{groups}{Character vector for groups as listed in dataset. Default = NA (no groups).}

\item{grlabs}{Character vector for group names, must be in same order as groups. Default = NA (no groups or use group levels from dataset).}

\item{state0}{Character label for state 0. Default = NA (do not display estimates for state 0).}

\item{ci.dec}{Number of decimal places to report for cumlative incidence estimates. Default = 2.}

\item{perc}{Logical indicator to report estimates as percentages. Default = FALSE.}

\item{color}{Character Hex color to use for htmlTable striping. Default = "#EEEEEE" (light grey).}

\item{printorig}{Logical indicator to print original summary of ms-survfit object for checking purposes. Default = TRUE.}
}
\description{
This function creates a summary table from a survfit object of type "mstate" including cumulative incidence estimates with 95% CIs.
It returns a dataframe and prints an html table by default for use in R markdown documents.
}
\keyword{Lisa}
\keyword{consulting}
\keyword{cumulative}
\keyword{incidence}
\keyword{mstate}
\keyword{summary}
\keyword{survival}
\keyword{table}
