% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurv.R
\name{ggsurv}
\alias{ggsurv}
\title{Lisa's Survival Plot}
\usage{
ggsurv(sfit, surv.lty = 1, surv.size = 1, surv.col = NA, ci = TRUE,
  ci.lty = 2, ci.size = 1, ci.ribbon = FALSE, cens = TRUE,
  cens.col = "red", cens.size = 1, cens.shape = 3, groups = NA,
  grlabs = NA, grname = NA, perc = FALSE, xlab = "\\n Time",
  ylab = "Survival \\n", main = "", themebw = TRUE)
}
\arguments{
\item{sfit}{survfit object (REQUIRED).}

\item{surv.lty}{Number for line type of survival lines. Default = 1.}

\item{surv.size}{Number for size of survival lines. Default = 1.}

\item{surv.col}{Character for color of survival lines. Default = "black".}

\item{ci}{Logical indicator to plot 95% confidence intervals. Default = TRUE.}

\item{ci.lty}{Number for line type of CI lines. Default = 2.}

\item{ci.size}{NUmber for size of CI lines, Default = 1.}

\item{ci.ribbon}{Logical indicator to use ribbons instead of lines for CI area. Default = FALSE.}

\item{cens}{Logical indicator to plot markers for censoring. Default = TRUE.}

\item{cens.col}{Character for color of censoring symbols. Default = "red".}

\item{cens.shape}{Number for censoring symbol (pch graphical parameter). Default = 3 (plus sign).}

\item{groups}{Character vector for groups as listed in dataset. Default = NA (no groups).}

\item{grlabs}{Character vector for group levels, must be in same order as groups. Default = NA (no groups or use group levels from dataset).}

\item{grname}{Character label for group legend name. Default = NA (no legend title).}

\item{perc}{Logical indicator to show y axis as percentages. Default = FALSE.}

\item{xlab}{Character label for x axis. Default = "Time".}

\item{ylab}{Character label for y axis. Default = "Survival".}

\item{main}{Character label for plot title. Default = " ".}

\item{themebw}{Indicator for use of ggplot2 theme_bw. Default = TRUE.}
}
\description{
This function creates a survival plot from a survfit object using ggplot2 graphics.
}
\keyword{KM}
\keyword{Lisa}
\keyword{consulting}
\keyword{curve}
\keyword{ggplot2}
\keyword{plot}
\keyword{summary}
\keyword{survival}
