% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicecoxph.R
\name{nicecoxph}
\alias{nicecoxph}
\title{Lisa's Cox Proportional Hazards Regression Table Reporting Function}
\usage{
nicecoxph(
  fit = NA,
  df = NA,
  covs = NA,
  ttevent = "ttevent",
  event = "event",
  ttentry = NA,
  weightvar = NA,
  cluster = NA,
  strata = NA,
  frailty = NA,
  regtype = "uni",
  type2 = FALSE,
  type3 = FALSE,
  phtest = FALSE,
  labels = NA,
  estname = NA,
  refcat = FALSE,
  est.dec = 2,
  ci.dec = 2,
  pval.dec = 3,
  color = "#EEEEEE",
  kable = TRUE,
  htmlTable = FALSE,
  title = "",
  footer = TRUE
)
}
\arguments{
\item{fit}{coxph model object [fit or df/event/ttevent/covs are REQUIRED].}

\item{df}{Dataframe [fit or df/event/ttevent/covs are REQUIRED].}

\item{covs}{Character. Vector of covariates to include in model [fit or df/event/ttevent/covs are REQUIRED].}

\item{ttevent}{Character. Name of time to event variable [fit or df/event/ttevent/covs are REQUIRED].}

\item{event}{Character. Name of event variable [fit or df/event/ttevent/covs are REQUIRED].}

\item{ttentry}{Character. Name of entry time variable for left truncated data. Default is NA (no delayed entry).}

\item{weightvar}{Character. Name of weighting variable. Default is NA (no weights).}

\item{cluster}{Character. Name of cluster variable. Default is NA (no clustering).}

\item{strata}{Character. Vector of names of strata variable(s). Default is NA (no strata).}

\item{frailty}{Character. Name of frailty variable. Default is NA (no random effect).}

\item{regtype}{Logical. Should the covariates be run separately ("uni") or together in a multiple regression model ("multi") [REQUIRED if no fit].}

\item{type2}{Logical. If TRUE, type II p-values will be added to the table. Default is FALSE.}

\item{type3}{Logical, If TRUE, type III p-values will be added to the table. Default is FALSE.}

\item{phtest}{Logical. If TRUE, a test for proportional hazards (via cox.zph) will be included in the footnote. Default is FALSE.}

\item{labels}{Character. Vector of labels for covariates. Default is NA (use variable names).}

\item{estname}{Character. Label for estimate. Default = NA: "HR" for univariate and "aHR" for multiple regression.}

\item{refcat}{Logical. If TRUE the table will create a separate line for the reference category. Default is FALSE.}

\item{est.dec}{Numeric. Number of decimal places for estimates. Default is 2.}

\item{ci.dec}{Numeric. Number of decimal places for confidence interval. Default is 2.}

\item{pval.dec}{Numeric. Number of decimal places for pvalues. Must be an integer from 1 to 4. Default is 3.}

\item{color}{Character. Color to use for htmlTable striping. Default is "#EEEEEE" (light grey). Use "white" for no striping.}

\item{kable}{Logical. Indicator to use kable to display table. Default = TRUE.}

\item{htmlTable}{Logical. Indicator to use htmlTable package to display table instead of kable Default = FALSE.}

\item{title}{Character. Optional title above table. Default is "".}

\item{footer}{Logical. If TRUE, table will include a footnote with model details, nobs, R2. Default is TRUE.}
}
\description{
This function creates a nice looking table of regression results for a coxph model.
Input either a coxph object or dataframe, event variable, time variable, vector of covariates, and type of analysis (univariate or multiple regression).
The function returns a dataframe of formatted regression results and prints the results table using kable or htmlTable.
}
\keyword{cox}
\keyword{coxph}
\keyword{hazards}
\keyword{proportional}
\keyword{regression}
\keyword{reporting}
\keyword{survival}
\keyword{table}
